fix Travis build scripts
- install newer setuptools if needed
- install transitive dependencies of ckanext-archiver and ckanext-report
- start to recognise Python 3

+++ bin/travis-build.bash
@@ -6,7 +6,16 @@ echo "This is travis-build.bash..."
 
 echo "Installing the packages that CKAN requires..."
 sudo apt-get update -qq
-sudo apt-get install solr-jetty libcommons-fileupload-java
+sudo apt-get install -y solr-jetty libcommons-fileupload-java
+
+ver=$(python -c"import sys; print(sys.version_info.major)")
+if [ $ver -eq 2 ]; then
+    echo "python version 2"
+elif [ $ver -eq 3 ]; then
+    echo "python version 3"
+else
+    echo "Unknown python version: $ver"
+fi
 
 echo "Upgrading libmagic for ckanext-qa..."
 # appears to upgrade it from 5.09-2 to 5.09-2ubuntu0.6 which seems to help the tests
@@ -15,6 +24,10 @@ sudo apt-get install libmagic1
 echo "Installing CKAN and its Python dependencies..."
 git clone https://github.com/ckan/ckan
 cd ckan
+
+if [ $ver -eq 3 ]; then
+    pip install -r requirement-setuptools.txt
+fi
 
 if [ $CKANVERSION == 'master' ]
 then
@@ -25,14 +38,14 @@ else
     echo "CKAN version: ${CKAN_TAG#ckan-}"
 fi
 
-python setup.py develop
-if [ -f requirements-py2.txt ]
+if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]
 then
     pip install -r requirements-py2.txt
 else
     pip install -r requirements.txt
 fi
 pip install -r dev-requirements.txt --allow-all-external
+python setup.py develop
 cd -
 
 echo "Setting up Solr..."
@@ -54,22 +67,29 @@ paster db init -c test-core.ini
 cd -
 
 echo "Installing dependency ckanext-report and its requirements..."
-pip install -e git+https://github.com/datagovuk/ckanext-report.git#egg=ckanext-report
-
-echo "Installing dependency ckanext-archiver and its requirements..."
-git clone https://github.com/ckan/ckanext-archiver.git
-cd ckanext-archiver
-pip install -e .
-pip install -r requirements.txt
+git clone --depth=50 https://github.com/datagovuk/ckanext-report.git
+cd ckanext-report
+  if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]; then
+    pip install -r requirements-py2.txt
+  elif [ -f requirements.txt ]; then
+    pip install -r requirements.txt
+  fi
+  pip install --no-deps -e .
 cd -
 
-echo "Installing ckanext-qa and its requirements..."
-python setup.py develop
-pip install -r requirements.txt
-pip install -r dev-requirements.txt
+echo "Installing dependency ckanext-archiver and its requirements..."
+git clone --depth=50 https://github.com/ckan/ckanext-archiver.git
+cd ckanext-archiver
+  if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]; then
+    pip install -r requirements-py2.txt
+  elif [ -f requirements.txt ]; then
+    pip install -r requirements.txt
+  fi
+  pip install --no-deps -e .
+cd -
 
 echo "Moving test-core.ini into a subdir..."
-mkdir subdir
-mv test-core.ini subdir
+mkdir -p subdir
+cp test-core.ini subdir
 
 echo "travis-build.bash is done."
